Minor cleanup and renaming files.
