+++ src/lib/propagation/constants.js
@@ -1,27 +1,3 @@
-// Types:
-//
-// Simulation Node:
-// {
-//   id: String,
-//   type: oneof(NODE_TYPES),
-//   expression: null if type === STYPE.data, else user expression.
-//   samples: [...],
-//   errors: [...],
-// }
-//
-// Global:
-// {
-//   id: String,
-//   samples: [...],
-// }
-//
-// DAG: [
-//   ...height 0 nodes: [{...simulation node, inputs: [...input node ids], children: [...indices of children w/in DAG]}],
-//   ...height 1 nodes,
-//   ...
-// ]
-
-
 export const NODE_TYPES = {
   UNSET: 0, // For safety; should not be used.
   DATA: 1,
@@ -30,12 +6,14 @@ export const NODE_TYPES = {
 }
 
 export const ERROR_TYPES = {
-  GRAPH_ERROR: 0,
+  UNSET: 0, // For safety; should not be used.
+  GRAPH_ERROR: 1,
 }
 export const ERROR_SUBTYPES = {
   GRAPH_SUBTYPES: {
-    MISSING_INPUT_ERROR: 0,
-    IN_INFINITE_LOOP: 1,
-    INVALID_ANCESTOR_ERROR: 2,
+    UNSET: 0, // For safety; should not be used.
+    MISSING_INPUT_ERROR: 1,
+    IN_INFINITE_LOOP: 2,
+    INVALID_ANCESTOR_ERROR: 3,
   },
 }
