+++ src/lib/propagation/nodeFns.js
